small int for rating to int

+++ green_creme/migrations/007_create_reply_table.py
@@ -7,7 +7,7 @@ steps = [
             forum_id int not null references forum (id),
             answer text not null,
             image varchar(400),
-            rating smallint not null default 0,
+            rating int not null default 0,
             created_on timestamp not null default current_timestamp,
             constraint reply_author_fk
                 foreign key (author_id) references accounts (id)
